employee_directory/api/permissions.py
from rest_framework.permissions import BasePermission


class IsEnrolled(BasePermission):
    def has_object_permission(self, request, view, obj):
        return obj.owner.filter(id=request.user.id).exists()
